backend/requirements.txt
# Web Framework
# Note: Django 4.1.13 is required due to djongo compatibility
Django==4.1.13
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1

# Database
mysqlclient==2.2.1

# Environment and Configuration
python-decouple==3.8
python-dotenv==1.0.0

# NLP and ML
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
transformers==4.55.3
tokenizers==0.21.4
torch==2.5.1

# API Documentation
drf-yasg==1.21.7

# Development & Testing
black==23.12.1
flake8==7.0.0
pytest==7.4.3
pytest-django==4.7.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# MongoDB and components
djongo==1.3.6
pymongo==4.5.0

# Additional dependencies that were missing
huggingface-hub==0.34.4
safetensors==0.6.2
regex==2025.7.34
requests==2.32.5

